A small code-clean and test run validation and minor correction in the view this ticket pertains to
Some new tests for the mock data.
And corrections in the project as a whole for Sonar.

Added tests for StatusTagHelper

Responding to SonarCube part 1

Improved coverage of RegulatorSubmissions Controller and the primary journey data model.

SonarCube dispose alert in a test....!

Dispose of In a Strict Mock cause a test failure

I forgot to delete a file duplicating code....