Add an AccessDenied page
The current behaviour on main is to 404 for access denied errors, which
in turn 500's.
Having fixed the 500'ing of 404s, we can now see the 404 for access
denied (when a user successfully signs in to b2c but doesn't have a
valid account in the backend sql).

By adding this controller-action and route we no longer get spurious
"not found" error shown to the user when it's an access problem.

The translation file is based on the text in the not found page as a
starting point.

https://eaflood.atlassian.net/browse/SMAL-205